+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/ReproducibleLayerBuilder.java
@@ -18,13 +18,19 @@ package com.google.cloud.tools.jib.image;
 
 import com.google.cloud.tools.jib.filesystem.DirectoryWalker;
 import com.google.cloud.tools.jib.tar.TarStreamBuilder;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Verify;
 import com.google.common.collect.ImmutableList;
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 
 /**
@@ -35,40 +41,79 @@ import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 public class ReproducibleLayerBuilder {
 
   /**
+   * Holds a list of {@link TarArchiveEntry}s with unique extraction paths. The list also includes
+   * all parent directories for each extraction path.
+   */
+  private static class UniqueTarArchiveEntries {
+
+    /**
+     * Uses the current directory to act as the file input to TarArchiveEntry (since all directories
+     * are treated the same in {@link TarArchiveEntry#TarArchiveEntry(File, String)}, except for
+     * modification time, which is wiped away in {@link #build}).
+     */
+    private static final File DIRECTORY_FILE = Paths.get(".").toFile();
+
+    private final List<TarArchiveEntry> entries = new ArrayList<>();
+    private final Set<String> names = new HashSet<>();
+
+    /**
+     * Adds a {@link TarArchiveEntry} if its extraction path does not exist yet. Also adds all of
+     * the parent directories on the extraction path.
+     *
+     * @param tarArchiveEntry the {@link TarArchiveEntry}
+     */
+    private void add(TarArchiveEntry tarArchiveEntry) {
+      if (names.contains(tarArchiveEntry.getName())) {
+        return;
+      }
+
+      // Adds all directories along extraction paths to explicitly set permissions for those
+      // directories.
+      Path namePath = Paths.get(tarArchiveEntry.getName());
+      if (namePath.getParent() != namePath.getRoot()) {
+        add(new TarArchiveEntry(DIRECTORY_FILE, namePath.getParent().toString()));
+      }
+
+      entries.add(tarArchiveEntry);
+      names.add(tarArchiveEntry.getName());
+    }
+
+    private List<TarArchiveEntry> getSortedEntries() {
+      List<TarArchiveEntry> sortedEntries = new ArrayList<>(entries);
+      sortedEntries.sort(Comparator.comparing(TarArchiveEntry::getName));
+      return sortedEntries;
+    }
+  }
+
+  /**
    * Builds the {@link TarArchiveEntry}s for adding this {@link LayerEntry} to a tarball archive.
    *
    * @return the list of {@link TarArchiveEntry}
    * @throws IOException if walking a source file that is a directory failed
    */
-  private static List<TarArchiveEntry> buildAsTarArchiveEntries(LayerEntry layerEntry)
-      throws IOException {
+  @VisibleForTesting
+  static List<TarArchiveEntry> buildAsTarArchiveEntries(LayerEntry layerEntry) throws IOException {
     List<TarArchiveEntry> tarArchiveEntries = new ArrayList<>();
 
+    // Adds the files to extract relative to the extraction path.
     for (Path sourceFile : layerEntry.getSourceFiles()) {
       if (Files.isDirectory(sourceFile)) {
         new DirectoryWalker(sourceFile)
             .filterRoot()
+            .filter(path -> !Files.isDirectory(path))
             .walk(
                 path -> {
-                  /*
-                   * Builds the same file path as in the source file for extraction. The iteration
-                   * is necessary because the path needs to be in Unix-style.
-                   */
-                  StringBuilder subExtractionPath =
-                      new StringBuilder(layerEntry.getExtractionPath());
                   Path sourceFileRelativePath = sourceFile.getParent().relativize(path);
-                  for (Path sourceFileRelativePathComponent : sourceFileRelativePath) {
-                    subExtractionPath.append('/').append(sourceFileRelativePathComponent);
-                  }
+                  Path extractionPath =
+                      Paths.get(layerEntry.getExtractionPath()).resolve(sourceFileRelativePath);
                   tarArchiveEntries.add(
-                      new TarArchiveEntry(path.toFile(), subExtractionPath.toString()));
+                      new TarArchiveEntry(path.toFile(), extractionPath.toString()));
                 });
-
       } else {
+        Path extractionPath =
+            Paths.get(layerEntry.getExtractionPath()).resolve(sourceFile.getFileName());
         TarArchiveEntry tarArchiveEntry =
-            new TarArchiveEntry(
-                sourceFile.toFile(),
-                layerEntry.getExtractionPath() + "/" + sourceFile.getFileName());
+            new TarArchiveEntry(sourceFile.toFile(), extractionPath.toString());
         tarArchiveEntries.add(tarArchiveEntry);
       }
     }
@@ -101,25 +146,35 @@ public class ReproducibleLayerBuilder {
    * @throws IOException if walking the source files fails
    */
   public UnwrittenLayer build() throws IOException {
-    List<TarArchiveEntry> filesystemEntries = new ArrayList<>();
+    UniqueTarArchiveEntries uniqueTarArchiveEntries = new UniqueTarArchiveEntries();
 
     // Adds all the layer entries as tar entries.
-    for (LayerEntry layerEntry : layerEntries.build()) {
-      filesystemEntries.addAll(buildAsTarArchiveEntries(layerEntry));
+    List<LayerEntry> layerEntries = this.layerEntries.build();
+    for (LayerEntry layerEntry : layerEntries) {
+      // Converts layerEntry to list of TarArchiveEntrys.
+      List<TarArchiveEntry> tarArchiveEntries = buildAsTarArchiveEntries(layerEntry);
+      // Adds the entries to uniqueTarArchiveEntries, which makes sure all entries are unique and
+      // adds parent directories for each extraction path.
+      tarArchiveEntries.forEach(uniqueTarArchiveEntries::add);
     }
 
-    // Sorts the entries by name.
-    filesystemEntries.sort(Comparator.comparing(TarArchiveEntry::getName));
+    // Gets the entries sorted by extraction path.
+    List<TarArchiveEntry> sortedFilesystemEntries = uniqueTarArchiveEntries.getSortedEntries();
+
+    Set<String> names = new HashSet<>();
 
     // Adds all the files to a tar stream.
     TarStreamBuilder tarStreamBuilder = new TarStreamBuilder();
-    for (TarArchiveEntry entry : filesystemEntries) {
+    for (TarArchiveEntry entry : sortedFilesystemEntries) {
       // Strips out all non-reproducible elements from tar archive entries.
       entry.setModTime(0);
       entry.setGroupId(0);
       entry.setUserId(0);
       entry.setUserName("");
       entry.setGroupName("");
+
+      Verify.verify(!names.contains(entry.getName()));
+      names.add(entry.getName());
 
       tarStreamBuilder.addTarArchiveEntry(entry);
     }