+++ jib-maven-plugin/CHANGELOG.md
@@ -7,10 +7,13 @@ All notable changes to this project will be documented in this file.
 
 - `<container><labels>` configuration parameter for configuring labels ([#751](https://github.com/GoogleContainerTools/jib/issues/751))
 - `<container><entrypoint>` configuration parameter to replace entrypoint ([#579](https://github.com/GoogleContainerTools/jib/issues/579))
+- `history` to layer metadata ([#875](https://github.com/GoogleContainerTools/jib/issues/875))
 
 ### Changed
 
 ### Fixed
+
+- Corrects permissions for directories in the container filesystem ([#772](https://github.com/GoogleContainerTools/jib/pull/772))
 
 ## 0.9.9